webapp/requirements.txt
flask==3.1.0
plotly==5.24.1
pandas==2.2.3
psycopg2-binary==2.9.10
gunicorn==23.0.0
